+++ libs/base_core/src/lang/ffi/c/libc.rs
@@ -1,0 +1,20 @@
+// devela_base_core::lang::ffi::c::libc
+//
+//!
+//
+
+use crate::{_TAG_NUM, _TAG_PRIMITIVE};
+
+#[doc = _TAG_PRIMITIVE!()]
+#[doc = _TAG_NUM!()]
+/// Equivalent to C’s `mode_t` type.
+///
+/// Numeric type used in POSIX interfaces to encode file modes and permission bits.
+pub type c_mode_t = u32;
+
+#[doc = _TAG_PRIMITIVE!()]
+#[doc = _TAG_NUM!()]
+/// Equivalent to C’s `off_t` type.
+///
+/// Signed integer type used by POSIX interfaces to represent file offsets and sizes.
+pub type c_off_t = i64;