+++ libs/base_core/src/lang/ffi/c/mod.rs
@@ -3,11 +3,13 @@
 #![doc = crate::_DOC_LANG_FFI_C!()]
 //
 
+mod libc; // c_mode_t, c_off_t
 mod reexports;
 
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
+            libc::*,
             reexports::*,
         };
     }